requirements.txt
requests>=2.28.1
click>=8.1.3
rich>=13.3.2
discord.py>=2.1.0